Refactor the main code to export an interface
  This is so packages can use the fancy pants Urban Dictonary lookup
  code from outside the main module.
