+++ .gitignore
@@ -43,3 +43,6 @@ Icon
 .Trashes
 
 udict
+
+# vim
+*swp
